refactor: 페이지 내에서 공통으로 사용되는 상수, 타입, 유틸 함수 분리 #57
