+++ frontend/src/pages/PlanPage/PlanPage.constants.ts
@@ -1,0 +1,9 @@
+export const TRAFFIC_ICON = {
+  SUBWAY: 'Subway',
+  BUS: 'Bus',
+} as const;
+
+export const TRAFFIC_TYPE = {
+  SUBWAY: '지하철',
+  BUS: '버스',
+} as const;
