+++ frontend/src/pages/PlanPage/PlanPage.types.ts
@@ -1,0 +1,3 @@
+import { TRAFFIC_ICON } from './PlanPage.constants';
+
+export type TrafficType = keyof typeof TRAFFIC_ICON;
