+++ frontend/src/pages/PlanPage/hooks/usePlanInfoQuery.ts
@@ -1,0 +1,13 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { getPlanInfo } from '@/api/plan';
+import { QUERY_KEY } from '@/constants/queryKey';
+
+const usePlanInfoQuery = () => {
+  return useQuery({
+    queryKey: [QUERY_KEY.planInfo],
+    queryFn: getPlanInfo,
+  });
+};
+
+export default usePlanInfoQuery;
